+++ app/models/node.ts
@@ -1,13 +1,11 @@
 import { attr, belongsTo, hasMany } from '@ember-decorators/data';
 import { computed } from '@ember-decorators/object';
 import { alias, bool, equal } from '@ember-decorators/object/computed';
-import EmberObject from '@ember/object';
 import { not } from '@ember/object/computed';
 import { htmlSafe } from '@ember/string';
 import { buildValidations, validator } from 'ember-cp-validations';
 import DS from 'ember-data';
 
-import { NodeLicense } from 'ember-osf-web/transforms/node-license';
 import defaultTo from 'ember-osf-web/utils/default-to';
 import getRelatedHref from 'ember-osf-web/utils/get-related-href';
 
@@ -66,6 +64,11 @@ export enum NodeType {
     Fork = 'fork',
     Generic = 'generic',
     Registration = 'registration',
+}
+
+export interface NodeLicense {
+    readonly copyrightHolders?: string;
+    readonly year?: string;
 }
 
 export default class NodeModel extends BaseFileItem.extend(Validations, CollectableValidations) {
@@ -265,10 +268,10 @@ export default class NodeModel extends BaseFileItem.extend(Validations, Collecta
             year = new Date().getUTCFullYear().toString(),
         } = (this.nodeLicense || {});
 
-        const nodeLicenseDefaults: NodeLicense = EmberObject.create({
+        const nodeLicenseDefaults: NodeLicense = {
             copyrightHolders,
             year,
-        });
+        };
 
         // Only set the required fields on nodeLicense
         const props = requiredFields.reduce(
@@ -276,7 +279,7 @@ export default class NodeModel extends BaseFileItem.extend(Validations, Collecta
             {},
         );
 
-        this.set('nodeLicense', EmberObject.create(props));
+        this.set('nodeLicense', props);
     }
 }